requirements-dev.txt
black~=23.7.0
isort~=5.12.0
pycodestyle~=2.10.0
pytest-django~=4.5.2
pytest-env~=1.0.1
